+++ apps/web/src/components/global/Markdown.tsx
@@ -1,0 +1,11 @@
+import { marked } from "marked";
+
+export const Markdown = ({ content }: { content?: string }) => {
+  return (
+    <div
+      dangerouslySetInnerHTML={{
+        __html: marked(content ?? ""),
+      }}
+    />
+  );
+};
