+++ apps/web/src/components/global/preferences/UserChip.tsx
@@ -16,10 +16,10 @@ import { useAsync, useSearchParam } from "react-use";
 import MdiAccountKey from "~icons/mdi/account-key";
 import MdiLogin from "~icons/mdi/login";
 import MdiLogout from "~icons/mdi/logout";
-import { supabase } from "../../models/supabase";
-import { useVersionTheme } from "../../utils/useVersionTheme";
-import { Logo } from "../global/Logo";
-import { ResponsiveDialog } from "../global/ResponsiveDialog";
+import { supabase } from "../../../models/supabase";
+import { useVersionTheme } from "../../../utils/useVersionTheme";
+import { Logo } from "../Logo";
+import { ResponsiveDialog } from "../ResponsiveDialog";
 
 const ThemedAuth: FC<{
   view?: ViewType;
