+++ apps/web/src/components/global/site-meta/About.tsx
@@ -7,8 +7,8 @@ import MdiGithub from "~icons/mdi/github";
 import MdiInformation from "~icons/mdi/information";
 import MdiTwitter from "~icons/mdi/twitter";
 import MdiWeb from "~icons/mdi/web";
-import { BUNDLE } from "../../utils/bundle";
-import { ResponsiveDialog } from "../global/ResponsiveDialog";
+import { BUNDLE } from "../../../utils/bundle";
+import { ResponsiveDialog } from "../ResponsiveDialog";
 
 const ExternalLink: FC<PropsWithChildren<{ href: string }>> = ({
   href,
