+++ apps/web/src/components/rating/io/ExportMenu.tsx
@@ -2,8 +2,8 @@ import { Button, Menu } from "@mui/material";
 import { FC, useId, useState } from "react";
 import { Entry } from "../../../pages/RatingCalculator";
 import { PlayEntry } from "../RatingCalculatorAddEntryForm";
-import { ExportToJSONMenuItem } from "./ExportToJSONMenuItem";
-import { RenderToOneShotImageMenuItem } from "./RenderToOneShotImageMenuItem";
+import { ExportToJSONMenuItem } from "./export/ExportToJSONMenuItem";
+import { RenderToOneShotImageMenuItem } from "./export/RenderToOneShotImageMenuItem";
 
 export const ExportMenu: FC<{
   entries: PlayEntry[];
