+++ apps/web/src/components/sheet/SheetDialogContent.tsx
@@ -28,7 +28,7 @@ import { calculateRating } from "../../utils/rating";
 import { DXRank } from "../global/DXRank";
 import { SheetDialogContentHeader } from "./SheetDialogContentHeader";
 import { SheetTitle } from "./SheetListItem";
-import { SheetTags } from "./SheetTags";
+import { SheetTags } from "./tags/SheetTags";
 
 const PRESET_ACHIEVEMENT_RATES = [
   100.5, 100, 99.5, 99, 98, 97, 94, 90, 80, 75, 70, 60, 50,
